+++ lib/persistence.dart
@@ -5,6 +5,7 @@ export 'src/persistence/dao/base_dao.dart';
 export 'src/persistence/dao/dao_field.dart';
 export 'src/persistence/dao/dao_type.dart';
 
+export 'src/persistence/query/expression.dart';
 export 'src/persistence/query/filter.dart';
 export 'src/persistence/query/sort.dart';
 export 'src/persistence/query/query_select.dart';